==> preprocessed data/new/1.REC_거래량_데이터_prophet.ipynb <==
import pandas as pd

df = pd.read_csv(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\REC 거래량\2017~2024.csv')

df

rename_cols = {
    'clsPrc' : '종가',
    'jejuOrdRecValue' : '제주 매도물량',
    'landLwPrc' : '육지 최저가',
    'trdCnt' : '매수 건수',
    'landTrdCnt' : '육지 체결건수',
    'landTrdRecValue' : '육지 체결물량',
    'landOrdRecValue' : '육지 매도물량',
    'jejuOrdCnt' : '제주 매도건수',
    'landOrdCnt' : '육지 매도건수',
    'landAvgPrc' : '육지 평균가',
    'totCnt' : '총 거래건수',
    'landUplmtPrc' : '육지 상한가',
    'bzDd' : '현물시장거래일',
    'trdRecValue' : '매수 물량',
    'landHgPrc' : '육지 최고가',
    'jejuTrdCnt' : '제주 체결건수',
    'landLwlmtPrc' : '육지 하한가',
    'jejuLwlmtPrc' : '제주 하한가',
    'jejuAvgPrc' : '제주 평균가',
    'totRecValue' : '총 거래물량',
    'bidTrdVal' : '거래금액',
    'jejuHgPrc' : '제주 최고가',
    'jejuTrdRecValue' : '제주 체결물량',
    'jejuUplmtPrc' : '제주 상한가',
    'rn' : '순번',
    'jejuLwPrc' : '제주 최저가'
}

df.rename(columns=rename_cols, inplace=True)

df

def remove_columns_jeju(df):
    #"제주"가 포함되지 않은 컬럼들만 선택
    filtered_df = df.loc[:,~df.columns.str.contains('제주')]
    return filtered_df

df = remove_columns_jeju(df)
df

df.info()

print(df['현물시장거래일'].head())

# int64 형식의 날짜를 문자열로 변환 후, pd.to_datetime 사용
df['현물시장거래일'] = pd.to_datetime(df['현물시장거래일'].astype(str), format='%Y%m%d')

# 변환 확인
print(df['현물시장거래일'].head())

df.columns

# 논문 기준 REC 거래 데이터에서 추출한 패턴 데이터


 - REC 거래 규모에 대한 Cyclic Pattern 지수

 - REC 체결 건수에 대한 Cyclic Pattern 지수

 - REC 체결 가격에 대한 Cyclic Pattern 지수

from prophet import Prophet
import pandas as pd

# 데이터 준비
df['현물시장거래일'] = pd.to_datetime(df['현물시장거래일'])

# 분석 대상 변수

# 거래 규모
ts_df = df[['현물시장거래일','육지 체결물량']].rename(columns={'현물시장거래일':'ds','육지 체결물량':'y'})

# 체결 건수
nc_df = df[['현물시장거래일','육지 체결건수']].rename(columns={'현물시장거래일':'ds','육지 체결건수':'y'})

# 체결 가격
ef_df = df[['현물시장거래일','육지 평균가']].rename(columns={'현물시장거래일':'ds','육지 평균가':'y'})

ts_df

ts_df.info()

# 거래 규모
ts_model = Prophet(yearly_seasonality=True, weekly_seasonality=True, daily_seasonality=False)
ts_model.fit(ts_df) 

# 체결 건수
nc_model = Prophet(yearly_seasonality=True, weekly_seasonality=True, daily_seasonality=False)
nc_model.fit(nc_df) 

# 체결 가격
ef_model = Prophet(yearly_seasonality=True, weekly_seasonality=True, daily_seasonality=False)
ef_model.fit(ef_df) 

# 미래 데이터 생성
future_ts = ts_model.make_future_dataframe(periods=365)
future_nc = nc_model.make_future_dataframe(periods=365)
future_ef = ef_model.make_future_dataframe(periods=365)

# 예측
forcast_ts = ts_model.predict(future_ts)
forcast_nc = nc_model.predict(future_nc)
forcast_ef = ef_model.predict(future_ef)

forcast_ts 

# 육지 체결 가격 트렌드 및 주기 성분 
df = df.merge(forcast_ef[['ds','trend','yearly']], left_on='현물시장거래일', right_on='ds', how='left')
df.rename(columns={'trend':'육지_체결_가격_트랜드', 'yearly':'육지_체결_가격_주기성분'}, inplace=True)

# 육지 체결 건수 트렌드 및 주기 성분
df = df.merge(forcast_nc[['ds','trend','yearly']], left_on='현물시장거래일', right_on='ds', how='left')
df.rename(columns={'trend':'육지_체결_건수_트랜드', 'yearly':'육지_체결_건수_주기성분'}, inplace=True)

# 육지 거래 규모 트렌드 및 주기 성분
df = df.merge(forcast_ts[['ds','trend','yearly']], left_on='현물시장거래일', right_on='ds', how='left')
df.rename(columns={'trend':'육지_거래_규모_트랜드', 'yearly':'육지_거래_규모_주기성분'}, inplace=True)

df.columns

df

df = df.drop(columns=['ds_x','ds_y','ds'], axis=1)

df

df.to_csv(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\전처리 데이터\dataframe_1.csv',index=False, encoding='utf-8')
