==> preprocessed data/new/2-1.RPS_의무량_병합_new.ipynb <==
import pandas as pd

df = pd.read_csv(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\전처리 데이터\dataframe_1.csv')

# df 불러온 뒤,
df['현물시장거래일'] = pd.to_datetime(df['현물시장거래일'])
df['년도'] = df['현물시장거래일'].dt.year

RPS = pd.read_csv(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\RPS 공급의무자별 의무공급량\이후 자료 (8년치)\combined_year.csv', encoding='cp949')

print(RPS)

RPS

# '의무공급량(MWh)' 값을 숫자로 변환 (쉼표 제거)
RPS['의무공급량(MWh)'] = RPS['의무공급량(MWh)'].replace(",","",regex=True).astype(int)

pivot_rec = RPS.pivot(
    index='년도',
    columns='구분',
    values='의무공급량(MWh)'
)

print(pivot_rec)

df.columns

pivot_rec.columns = [str(col) + '_RPS' for col in pivot_rec.columns]

pivot_rec

# df의 현물시장거래일로부터 년도 추출
df['현물시장거래일'] = pd.to_datetime(df['현물시장거래일'])
df['년도'] = df['현물시장거래일'].dt.year

# pivot_mwh의 인덱스 -> 컬럼
pivot_mwh = pivot_rec.reset_index()

df_merged = pd.merge(
    df,
    pivot_mwh,
    how='left',  # 필요에 따라 'inner' 등 변경
    on='년도'
)

df_merged

df_merged.to_csv(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\전처리 데이터_수량증가\new_dataframe_2.csv',index=False,encoding='cp949')
