==> preprocessed data/new/3-1.SMP_데이터_병합.ipynb <==
import pandas as pd

df = pd.read_csv(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\전처리 데이터_수량증가\new_dataframe_2.csv',encoding='cp949')

smp_2017 = pd.read_excel(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\SMP\smp_land_2017.xlsx')
smp_2018 = pd.read_excel(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\SMP\smp_land_2018.xlsx')
smp_2019 = pd.read_excel(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\SMP\smp_land_2019.xlsx')
smp_2020 = pd.read_excel(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\SMP\smp_land_2020.xlsx')
smp_2021 = pd.read_excel(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\SMP\smp_land_2021.xlsx')
smp_2022 = pd.read_excel(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\SMP\smp_land_2022.xlsx')
smp_2023 = pd.read_excel(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\SMP\smp_land_2023.xlsx')
smp_2024 = pd.read_excel(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\REC 예측 관련데이터\SMP\smp_land_2024.xlsx')

# 데이터프레임의 0번 인덱스를 컬럼으로 설정하는 함수
def set_first_row_as_columns(df):
    df.columns = df.iloc[0]  # 0번 인덱스를 컬럼으로 설정
    df = df[1:].reset_index(drop=True)  # 0번 인덱스 데이터 제거 후 인덱스 리셋
    return df

# 데이터프레임 리스트
dataframes = [smp_2017,smp_2018,smp_2019,smp_2020,smp_2021,smp_2022, smp_2023, smp_2024]

# 각 데이터프레임에 함수 적용
smp_2017,smp_2018,smp_2019,smp_2020, smp_2021, smp_2022, smp_2023, smp_2024 = [set_first_row_as_columns(df) for df in dataframes]

smp_2020.head()

smp_combined = pd.concat([smp_2017,smp_2018,smp_2019,smp_2020, smp_2021, smp_2022, smp_2023, smp_2024], ignore_index=True)
smp_combined

# 최소, 최대, 평균 컬럼 이름 변경
smp_combined.rename(columns={
    '최소': 'SMP_최소',
    '최대': 'SMP_최대',
    '평균': 'SMP_평균',
    '구분' : '현물시장거래일'
}, inplace=True)

print(smp_combined.columns)

# int64 형식의 날짜를 문자열로 변환 후, pd.to_datetime 사용
smp_combined['현물시장거래일'] = pd.to_datetime(smp_combined['현물시장거래일'].astype(str), format='%Y%m%d')

df['현물시장거래일'] = pd.to_datetime(df['현물시장거래일'])

# 필요한 컬럼만 선택 (현물시장거래일, SMP_최소, SMP_최대, SMP_평균)
smp_combined_subset = smp_combined[['현물시장거래일', 'SMP_최소', 'SMP_최대', 'SMP_평균']]

# df와 smp_combined 병합 (현물시장거래일 기준)
df = pd.merge(df, smp_combined_subset, on='현물시장거래일', how='left')

df.head()

# NaN 값이 하나라도 있는지 확인
has_nan = df.isnull().values.any()

print(f"NaN 값이 존재합니다: {has_nan}")

df.to_csv(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\전처리 데이터_수량증가\new_dataframe_3.csv',index=False,encoding='cp949')
