==> preprocessed data/new/4-1.RPS의무비율_물가상승률_경제성장률_병합.ipynb <==
import pandas as pd

df = pd.read_csv(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\전처리 데이터_수량증가\new_dataframe_3.csv',encoding='cp949')

df

# RPS 의무비율 
## 출처 - 한국에너지공단 신재생에너지센터
 - 2017 / 4.0%
 - 2018 / 5.0%
 - 2019 / 6.5%
 - 2020 / 7.0%
 - 2021 / 9.0%
 - 2022 / 12.5%
 - 2023 / 13%
 - 2024 / 13.5%

test= df

import numpy as np

# 조건 정의
conditions = [
    test['년도'] == 2017,
    test['년도'] == 2018,
    test['년도'] == 2019,
    test['년도'] == 2020,
    test['년도'] == 2021,
    test['년도'] == 2022,
    test['년도'] == 2023,
    test['년도'] == 2024
]

# 각 조건에 할당할 값 정의 
values = [4.0,5.0,6.5,7.0, 9.0, 12.5, 13, 13.5]

# 의무 RPS 비율 컬럼 생성 및 값 할당
test['의무RPS비율'] = np.select(conditions, values, default=np.nan)

test.head()

## 경제 성장률 
### 출처 - 한국은행
 - 2017 / 3.4%
 - 2018 / 3.2%
 - 2019 / 2.3%
 - 2020 / -0.7%
 - 2021 / 4.6%
 - 2022 / 2.6%
 - 2023 / 1.4%
 - 2024 / 2.4%

import numpy as np

# 조건 정의
conditions = [
    test['년도'] == 2017,
    test['년도'] == 2018,
    test['년도'] == 2019,
    test['년도'] == 2020,
    test['년도'] == 2021,
    test['년도'] == 2022,
    test['년도'] == 2023,
    test['년도'] == 2024
]

# 각 조건에 할당할 값 정의 
values = [3.4,3.2,2.3,-0.7, 4.6, 2.6, 1.4, 2.4]

# 의무 RPS 비율 컬럼 생성 및 값 할당
test['경제성장률'] = np.select(conditions, values, default=np.nan)

test.head()

# 물가지수 확인

## 출처- 통계청
## 2024년은 11월 기준으로 소비자 물가동향 설정
### 2020년을 100으로 기준잡아서 물가동향을 계산함
 - 2017 / 97.6
 - 2018 / 99.1
 - 2019 / 99.5
 - 2020 / 100.00
 - 2021 / 102.50
 - 2022 / 107.72
 - 2023 / 111.59
 - 2024 / 114.40

# 조건 정의
conditions = [
    test['년도'] == 2017,
    test['년도'] == 2018,
    test['년도'] == 2019,
    test['년도'] == 2020,
    test['년도'] == 2021,
    test['년도'] == 2022,
    test['년도'] == 2023,
    test['년도'] == 2024
]

# 각 조건에 할당할 값 정의 
values = [97.6,99.1,99.5,100.00, 102.50, 107.72, 111.59, 114.40]

# 의무 RPS 비율 컬럼 생성 및 값 할당
test['물가지수'] = np.select(conditions, values, default=np.nan)

test.head()

test.to_csv(r'C:\Users\PC027\Desktop\Project\KOSPO공모전\전처리 데이터_수량증가\new_dataframe_4.csv',index=False,encoding='cp949')
